shocktube_y is now succeeded. the key is the compiler option. use double precision and fast_math, m32 explicitly conquer shocktube problem

+++ bphcuda/distribution/shell_distribution.h
@@ -31,7 +31,8 @@ public:
   __host__ __device__
   real3 operator()(const real2 &rand) const {
     real cs = real(1) - real(2) * rand.get<0>(); // cs = [-1, 1)
-    real sn = sqrt(real(1) - cs*cs);
+    // real sn = sqrt(real(1) - cs*cs);
+    real sn = sqrtf(real(1) - cs*cs);
     real b = real(2) * _PI * rand.get<1>();
     real cx = sn * sinf(b);
     real cy = sn * cosf(b);
